requirements.txt
boto3==1.20.0
kafka-python==2.0.2
klingon-file-manager
